Fix v60001/concbkup subtest failure on Arch linux due to a tcsh 6.20 issue
The v60001/concbkup subtest does the following to filter out BKUPRUNNING and MUNOFINISH messages from $errlog

	alias knownerror 'mv \!:1 {\!:1}x ; $grep -vE "\!:2" {\!:1}x >&! \!:1 '
	knownerror $errlog "BKUPRUNNING\|MUNOFINISH"

But this fails as follows.

	grep: Trailing backslash

This is because tcsh 6.20 splits the \| (before the MUNOFINISH) as a \ followed by a | (command pipeline)
and so splits the command

	knownerror $errlog "BKUPRUNNING\|MUNOFINISH"

as a pipeline sequence of the below two commands.

	knownerror $errlog "BKUPRUNNING\
	MUNOFINISH"

And the grep inside the knownerror alias issues the trailing backslash error because of this.
The issue is with tcsh 6.20.

tcsh 6.18 does not have this issue. But instead of downgrading tcsh, it was considered better to avoid such issues
in the test system. And that is now done by replacing the alias knownerrors with a $gtm_tst/com/knownerrors.csh
script. And passing in the search string as "BKUPRUNNING|MUNOFINISH" (without the backslash). i.e.

$gtm_tst/com/knownerror.csh "BKUPRUNNING|MUNOFINISH"

This avoids the issue and lets the test pass on platforms irrespective of the tcsh version.
The knownerrors alias was defined in lots of tests. This commit fixes all of them to instead use the new script
and as a side effect also avoids the duplicate alias definition throughout the test scripts.