[YDB#12] [NARS] [estess] Fix v53003/C9I09003044 reference file now that clue is maintained in more places than before
Test
-----
* E_ALL run many times to ensure no regressions.

README
-------
For the BG access method, the CAT (successful critical section acquisitions) statistic was zero
previously but after the code changes for YDB#12, it is not. This is because the test1f section
of the v53003/C9I09003044 subtest tests for the existence of a non-existent global node inside
a TP transaction. It does 8 types of operations each inside a separate TP which causes the
CAT value to be 8 at the end whereas previously it used to be 0. Searching for a non-existent
global means scanning the directory tree for this global name. With the code changes to
maintain clue in more cases (see https://github.com/YottaDB/YottaDB/pull/29 for more details), the
directory tree clue is now maintained whereas it previously was not. But because a lot of unrelated
updates happened in this database since this clue was established, a read-only fast-path optimization
in tp_tend ("si->start_tn == csd->trans_hist.early_tn") did not kick in like before which meant we
now validate this TP transaction by grabbing crit (call to "grab_crit") which causes the CAT value
to be higher. So we save on block traversals inside the directory tree due to the clue but we now
incur the cost of a grab_crit at commit time for read-only transactions. Note that searches for
globals that exist do not suffer from this issue since once a global name is found in the directory
tree, that tree is never searched again even if that global is later killed. Non-TP (t_end) has a
fast-path for read-only transactions that validates the history outside of crit but TP (tp_tend)
currently does not have such a scheme. If/when that is done, we will have the benefit of the clue
as well as avoid the "grab_crit" call.

For the MM access method, the DRD (Disk ReaDs) statistic reduced due to the clue enhancement. The
reference file was fixed accordingly.