Fix v63000/gtm6388 test failure due to strace output format differences between Ubuntu and Arch Linux
The test was trying to find out the fd number in the "open" system call. And checking the # of write()
system calls using this fd. In Ubuntu the system call was "open" but in Arch Linux it was "openat".
Below is the actual strace output.

Ubuntu 16.04 : open("extract.ZWR", O_RDWR|O_CREAT|O_NOCTTY, 0664) = 4
Arch Linux   : openat(AT_FDCWD, "/extra3/testarea1/nars/V998/tst_V63002_R100_dbg_02_170922_111941/v63000_0_1/gtm6388/mumps.gld", O_RDONLY) = 3

Because of this, the test (which was using the pattern "open..extract") to match, only found a matching
line in the Ubuntu output but not in the Arch output.

This is now fixed to search for the pattern "open.*extract" that matches both the above lines.