Do not start a test if $ggdata and/or subdirectories are not writeable
Previously, the test got submitted but failed after a while when trying to log the test results etc. in case the NFS directory corresponding to $ggdata was not mounted. It is better to know of this issue at the start of test submission (i.e. no tests have started) so the NFS issue can be fixed and the test resubmitted.