Get java test to actually run (and pass) in YottaDB setup
* com/set_java_paths.csh : If $gtm_test_serverconf_file does not define an explicit location for the java installation,
try to find it out on-the-fly. This will avoid having to hardcode the java path for each new system added to the
testing cluster.

* com/set_java_supported.csh : Do not redirect output of set_java_paths.csh call to /dev/null as it might be useful
in debugging.

* com/submit.csh : Now that set_java_paths.csh call in set_java_supported.csh does not redirect its output to
/dev/null, the parent script invocation (set_java_suported.csh) could produce a non-zero output so redirect that
to a file in $gtm_test_local_debugdir

* java/inref/com.test.ji.TestCase.java : gtmji is now installed in $gtm_dist/plugin (and not
/usr/library/com). i.e. the plugin is version specific.

* java/instream.csh : setenv gtmji_dir to $gtm_dist/plugin as that is the new gtmji install path; Run the test
always, whether or not java is supported on this platform (previously the test would pass indicating the java test
actually ran even though it did not run any subtest)