Fix java/callins subtest failure due to change in HALT behavior inside a call-in (in r1.10, see YottaDB/YottaDB#32)
As part of r1.10, a code change was made so HALT inside a call-in returns back to the C or Java caller instead of exiting the process. This caused Test 9 to behave differently. The expected output as well as the actual output (both of which are generated) are now modified to reflect new code behavior.