Fix v61000/gtm7756_1 failure due to default GDS block size changing from 1K to 4K in V63002
Below is the test diff.

gtm7756_1.diff
--------------
6a7
> TEST-E-FAIL: updates count = 1999

The test script (gtm7756_1.csh) has a magic number of updates (1993) which is used to decide
which one of two possible messages (MUTRUNCNOSPACE or "Truncated region") to expect when a
MUPIP REORG -TRUNCATE is run. This number was derived at a time when the default GDS block
size was 1Kb. That got bumped to 4Kb in GT.M V6.3-002. That changed the magic number to a much
higher number (almost 4 times bigger). And that in turn caused the test to fail when it had
randomly chosen the # of updates to be > 1993 (in this case it was 1999). Instead of experimenting
and finding out the new number, the GDS block size is explicitly set to 1Kb in the test and that
should keep the magic number of 1993 accurate and let this test pass for all random # of updates
chosen in gtm7756.m ("updatecnt" variable).