Fix v63000/gtm6388 subtest failure due to gawk 4.2.0 change of behavior in handling +$2 where $2 contained a leading space
Prior to gawk 4.2.0, if $2 contained " 4 <0.000017>", a +$2 returned the numeric value of "4".
But in gawk 4.2.0, it returns " 4" and that causes a test failure since the test is then searching for
occurrences of write( 4,...) in the strace output instead of looking for write(4,...) in the strace output.
It is not clear whether this change is a regression or a permanent change of behavior in gawk.
Hence this is now fixed by using sub() command and avoiding the +$2 operator altogether.