Fix 64bittn/rolling_upgrade test hang in case DBDANGER is seen after crash on receiver side
In one test run, this test did a crash of the receiver side (kill -9) without removing shared memory and then
did a rundown (as part of mupip_rollback.csh call in shut_normal_kill.csh) which encountered a DBDANGER error
(possible in case of kill -9 situations) and that in turn froze the instance (because this test randomly ran with
custom errors turned on). And that caused the test to hang indefinitely since there was no one there to manually
unfreeze the instance.

In this case, the receiver side crash should also have removed shared memory. Leaving shm around is what caused the
rundown to try to flush it and resulted in the DBDANGER. Any test that does crashes removes shared memory in the
test system. If it does not, that is then a special test which is run manually (e.g. manually_start test). Since this
test is better kept as part of the regular E_ALL, the NO_IPCRM usage in the receiver_crash.csh call has been removed.