[#94] Fix v62000/gtm7926rcvr subtest failure; Add debug information to v62000/gtm8047 subtest
1) The v62000/gtm7926rcvr subtest failed due to the recent YottaDB/YottaDB#94 changes
(ydb_dist/gtm_dist env var handling changes at process startup). It is now fixed to expect
the receiver server and udpate process startup to succeed even if gtm_dist/ydb_dist is not
set correctly at time of startup of mupip executable.

Took this opportunity to additionally enhance the test coverage by

* Choosing ANY prior version instead of a specific range like before.
* Enabling the test to be run in DBG as well (not just PRO).

2) The v62000/gtm8047 subtest limits the virtual memory available to a process and then
execute a TP transaction that never ends. We eventually expect a MEMORY error or various
forms of out-of-memory error strings. But no such output shows up once in a while resulting
in a test failure. There is zero output in gtm8047.outx. A change has therefore been made to
the M program (gtm8047.m) so it prints each M line as it executes it. That way we have at
least some output if the M program did get control. If a failure happens after this change
and there is zero output, then it is definitely the case the OS has terminated the process
abruptly at startup and that need to separately handled then to avoid false failures.