Fix gtcm_gnp tests (-gtcm) to work with ipv6 named hosts (e.g. $host.v46)
dbcreate.csh in case of a test started with -gtcm option (all the _16 tests)
invokes dbcreate_base.csh on all the hosts (client and 2 servers). If ipv6
support is detected, the test framework randomly sets the server hostname
to be ${host}.v46 or ${host}.v6 or ${host}.v4 (instead of ${host}). In this
case, dbcreate.csh on the client invokes dbcreate_base.csh on the server using
the ${host}.v46 value i.e. "dbcreate_base.csh -GTCM_REMOTE ${host}.v46:<path>"
where <path> is the path where the database files need to be created. In this
case though, dbcreate_base.csh uses an awk script dbcreate_multi.awk that tries
to match ${host}.v46 against the current host (one of the servers) and if it
does not match, it skips creating the database on this server because it think
this database is destined for the other server. This causes no databases like
b.dat, c.dat etc. to be created on the servers.

This is now fixed by passing only ${host} to dbcreate_base.csh (removing the
.v46, .v4 or .v6 hostname suffix) and that results in the appropriate databases
being created.