Allow for wider range of <max node len> due to enhanced testing of random sprgde files in E_ALL
There was a test failure in the collation/pol_to_revpol test. A max node len of 35 or
36 or 37 was accepted in the reference file but we ended up with 33. This is possible
because max node len in turn translates to the maximum record size (the maximum of the
"rsiz" field in the record header of the many records a GDS block has) and this rsiz
field can change because of a different distribution of the same set of nodes across
the available regions (due to the sprgde file mapping randomization, a recent test
change to generate .sprgde files randomly).

All reference files that already had a small allowance have now been fixed to widen the
range of allowed values. In the specific failure case, 3[567] was changed to 3. in the
reference file i.e. any of the values 30, 31, 32, ... , 39 are now considered acceptable.