[#59] Fix test timing issue in simpleapi/tp that can show up on slower systems
On a slow system it is possible that the first iteration of the "for" loop (based on
"iters" variable) takes more than 15 seconds to finish. In that case the break happens
in the first iteration of the "for" loop while "iters" is still 0 which results in the
"if (iters && cumulincrs)" check to fail and the "PASS from tp2" line to not be printed
resulting in a test failure. In this case, "cumulincrs" is guaranteed to be non-zero so
the check is now fixed to be "if (cumulincrs)". This fixes the test timing issue.