[V63004] Fix forw_rollback/verify test failure due to GT.M V6.3-004 code merge
GTM-6747 changed MUPIP SET -JOURNAL -REGION "*" behavior by creating new journal
files with the sequence number of the corresponding region (instead of using the
maximum of the region sequence numbers). This meant that the db and jnl seqnos
of a.mjl/a.dat were now equal in this test whereas they used to be different
and resulted in a JNLDBSEQNOMATCH error before.

So the test script was changed to choose b.mjl for the forward rollback as b.mjl/b.dat
were still out of sync in this test. But before printing a JNLDBSEQNOMATCH error,
a JNLDBTNNOMATCH error would be issued (since if seqnos are out of sync, tns will also
be out of sync). Therefore the test script was enhanced to first expect a JNLDBTNNOMATCH
error and then fix the database current transaction number of b.dat so it matches the
beginning transaction number of b.mjl and the forward rollback redone. This resulted
in the JNLDBSEQNOMATCH error (an error condition this test expects to exercise).