Fix rare/random simpleapi/nodenext subtest failure with MINNRSUBSCRIPTS error
Once in a while the nodenext subtest fails with the following error.

ydb_node_next() failure with rc -151027786 and error: 151027786,(SimpleAPI),
	%YDB-E-MINNRSUBSCRIPTS, Number of subscripts cannot be a negative number

This turns out to be due to an uninitialized variable "subs_out_cnt" which can
end up taking on a negative value. For the first iteration of the for loop
(i.e. when "reccnt" is 0), the value of "subs_in_cnt" is derived from "subs_out_cnt"
and in case "subs_out_cnt" is negative (possible in rare cases for uninitialized
variables), "subs_in_cnt" will end up being negative before being passed to the
"ydb_node_next_s()" call which can in turn cause the MINNRSUBSCRIPTS error return.