Fix timing issue in suppl_inst_C/gtm8074 subtest
It is possible, in rare cases, when we shut down the P->Q source server that P
is ahead of Q (i.e. all of the updates on P have not yet been replicated to Q).
In that case, since the later parts of the test do not do anything with Q, the
dbcheck at the end of the test will notice Q as lagging P and in turn lagging A.
This commit fixes that issue by waiting for P->Q to be in sync before shutting
down the corresponding source server.