Fix occasional v60002/gtm6548a and v60002/gtm6548c subtest failures with TEST-E-LSOF
The LSOF error was because jobbed off processes processes started inside the test were still running when the subtest driver script was done and handed control back to the test framework. The fix is to wait for these jobbed off processes to terminate before returning from the M program that does the jobbing.