Fix v60001/tprestart subtest failures due to lack of bitmap restarts (code: a)
In one test run, the last stage of this subtest did not encounter even one bitmap
related restart (code 'a'). In the syslog, a lot of G type restarts were seen
which are indicative of lack of global buffers. So the test is enhanced to have
8 times the default # of global buffers (from 1Ki to 8Ki). That should almost
eliminate the likelihood of G type restarts and increase the likelihood of 'a'
type restarts which this test expects.

While at that, also increased the allocation and extension settings 4 times
from 2Ki to 8Ki to reduce the frequency of DBFILEXT syslog messages. This way
the syslog will most likely have just the 'a' type restarts this test cares about.