Test JNLSWITCHRETRY error handling in source server (tests YottaDB/YottaDB#235)
This is done by enhancing the v62000/gtm8086 test (which is where the original failure
of #235 was seen) so the JNLSWITCHRETRY scenario is more frequently exercised
(50% of the time now compared to < 1% without these test enhancements).

The enhanced test failed reliably with JNLSWITCHRETRY error in the source server log
(in pro) and an assert failure (in dbg) when run with a version of the code that did
not have the #235 code fixes (and when "jnlswitchretry" variable was set to 1 which was
approximately 50% of the time). And passed reliably when run with the code fixes.