Fix test failures with rare failure symptom <TEST-E-PORTNO : Invalid port number>
There was an in-house test failure with the below failure symptom
	TEST-E-PORTNO : Invalid port number : head: cannot open ¿/tmp/test_36322.tx
		¿ for reading: No such file or directory 3607

In this case, the "head" command on a port reservation file resulted in an error because
the reservation file was concurrently deleted (by some other test that was in the
process of releasing that port number) and the error text from "head" (which was of
the form "cannot open ... for reading") got placed in the final portno choice of 36507.
So instead of portno returned from portno_acquire.csh being just "36507" it ended up being
"head: cannot open '/tmp/test_36322.txt' for reading: No such file or directory 36507".

The fix is to redirect the stderr (in addition to stdout) of the "head" command into
the pipeline that is being fed to the "grep" command, i.e. | changed to a |&