Increase hangalert time to 10 hours on 1-CPU systems to avoid TEST-E-HANG failures
The msreplic_H_1/max_connections subtest took > 6 hours to run in an in-house test run
on an ARMV6L box. This was while a system backup was already pounding on the disk.
And the test framework allowed for only a 5-hour max timeout. So it issued a
TEST-E-HANG message in the hangalert.out file which eventually caused the non-zero
diff file. Other than this TEST-E-HANG message, there was no diff meaning the test
passed otherwise.

This commit tries to prevent such failures by allowing for a bigger hangalert max
timeout (10 hours instead of 5 hours) for 1-CPU systems.