Fix potential TEST-E-TIMEOUT failure on slow systems waiting for REORG.TXT* to be removed
On a slow system where a MUPIP SIZE operation takes more than half an hour, it is possible
the the backgrounded online_reorg.csh script takes more than 1-1/2 hours to finish the
3 MUPIP SIZE operations (in case it chose MUPIP SIZE randomly) before recognizing that
the close_reorg.csh script invoked by the foreground test script had asked for it to stop
running. This could cause the close_reorg.csh script to time out as it has a timeout of
5400 seconds (1-1/2 hours) currently.

The fix is to rework the online_reorg.csh script so it does the check of the stop signal
between each of the 3 MUPIP SIZE operations instead of after all the 3. This should
reduce the maximum time that it can take for the pending stop signal to be seen.

While at this, a few other changes were done.

a) I noticed an issue in the script where it overwrote the online_reorg_*.outx file instead
   of appending to it in one place. This would cause prior rounds of the reorg output to be
   erased from the log file and make analysis of a test failure difficult.

b) Enhanced the script to record what choice it made and the actual MUPIP SIZE or MUPIP
   REORG commands it ran.