Fix mem_stress_1/memleak subtest failures/hangs due to system swapping
If a system has a few Mb of MemFree available but many Gb of SwapFree available,
it is possible currently for tst_buffsize to be set to 1/2Gb. This would cause
most of the journal and receive pools (3 of those each of 1/2Gb) to take a total
of 1-1/2Gb of space and most of it in the swap file. This can slow things down
terribly as was seen in an in-house failure/hang in an armv7l box.

Therefore, avoid the use of swap space and just base calculations on MemFree.