Bump RF_sync.csh timeout from 3 hours to 12 hours (to prevent test failures on slow systems)
The 3 hour timeout was derived from the $test_timeout env var which was used only in one
place in the entire test system so all usages of this are now removed and instead a
hardcoded constant used instead in that one place.

While at this, com/wait_until_src_backlog_below.csh was enhanced so all usages of writes to
the logfile have been changed to append that way we don't inadvertently overwrite content
in the logfile that pre-existed before entry into this script.