New v63004/gtm8914 subtest (tests GTM-8914 in V63004)
tests the following release note:

$VIEW("GVSTATS",<region>), ZSHOW "G" and ZSHOW "T"
return a question-mark (?) at the end of their output strings when the process does not have access to the current shared statistics;
they did not do this previously. Note that it is possible, although unlikely, the flag character could disrupt code that parses the results of these facilities.
(GTM-8914) [!Alert!] [!New Feature!]