Enhance submit_test.csh to send failure email for various cases when the test could not run
In in-house testing, we had a test failure where the YottaDB version used for testing
did not exist on a remote machine in a GT.CM test. But no failure mail was sent even
though there was a diff.log generated. This makes it harder to know the list of failures.

submit_test.csh has now been enhanced to identify all possible non-zero exit status and
ensure we signal a failure in all those cases. As well as print as much information as
possible to the user in the terminal where the test was submitted (as well as in the
failure email that gets sent) describing the reason for the failure.