Fix armv6l test failures in tests that relied on ydb_procstuckexec scheme
This feature was disabled at the test framework level on armv6l boxes in a recent commit.
The disabling was done to avoid the gdb tracing to slow down an already slow-running
heavyweight test. But there are a few simple tests that rely on this feature and those
can be enabled without any risk of failures.