New v63004/gtm8643 subtest (tests GTM-8643 in V63004)
Tests the following release note

The listen queue depth specified with WRITE /LISTEN for a listening socket is limited only by the limit imposed by the OS; previously, GT.M enforced an artificial limit of 5.

The below was also part of the same release note but is not tested as it is not straightforward to test this.

A SOCKET device retries connection attempts on local sockets that fail due to possible transient issues on the other end of the attempted connection, (for example an insufficiently large listen queue). Previously such attempts could sometimes appear to succeed, when they actually did not, leading to subsequent errors from READ and WRITE on the socket. (GTM-8643)