Fix occasional reorg/on_ntp_njnl_reorg subtest failure (NOSELECT error from MUPIP SIZE)
on_ntp_njnl_reorg.csh invokes enable_nct.csh which does a set^%GBLDEF of ^a,^b, etc.
That M routines sets numeric collation for global names by a two step process. A SET ^a
first followed by a KILL ^a etc. It is possible that online_reorg.csh (a script that runs
in the background, finds ^a existing if it runs mumps -run waitforglobal just after the
SET ^a in set^%GBLDEF but before the KILL ^a. This will cause online_reorg.csh to
incorrectly assume that at least one global is guaranteed to exist in the database
and proceed with MUPIP SIZE and/or MUPIP REORG operations which could then fail with a
YDB-W-NOSELECT (and YDB-E-MUNOACTION error).

The fix is to remove the call to enable_nct.csh in reorg/u_inref/on_ntp_njnl_reorg.csh
since the global names manipulated in that script are not anyways used by this subtest.