Enhance ext_filter_1/notalive subtest to also test YottaDB/YottaDB#264
The source server loops indefinitely with r1.22 x86_64 and arm linux with this revised
testcase. But passes with the code fix for #264. Whereas the previous test case used
to fail rarely only on arm linux and pass always on x86_64 linux.

The enhanced test case (which includes a hang in notalive.m) is done randomly so
we exercise the previous test codepath too.