Fix v62000_1/gtm8023 subtest failure on slow systems by bumping timeout from 10 minutes to 1 hour
On an armv6l box, we had a test failure where readonly^gtm8023 timed out in 10 minutes
after waiting for ^nowprimary to get created. That node did get created in 15 minutes.
It was just a slow system (Raspberry Pi Zero armv6l boxes are 1-CPU).
The fix is to bump the timeout from 10 minutes to 1 hour.