New v63004/gtm8903 subtest (tests GTM-8903 in V63004) (#190)
Tests the following release note

$SELECT() deals with cases where the first value is evaluated to a literal TRUE (1) and later arguments to the function
contain one or more global references. Due to a regression in V6.3-002 associated with compiler optimizations
caused this combination to produce a SIG-11 (segmentation violation) at compilation. (GTM-8903)

NOTE:
With the information provided in the patch notes, a sig-11 could not be reproduced. The test instead is designed to produce a GTMASSERT2 in regressed releases.