Fix v63003/gtm8735 subtest failure on systems where <realpath $ydb_dist> is different from $ydb_dist
We had a failure of the v63003/gtm8735 subtest on a host where $ydb_dist was
/usr/library/V999 but "realpath $ydb_dist" was "/extra/usr/library/V999". That is the
unexpanded and expanded file paths were different. This confused the test framework
when the test output had a mix of the unexpanded and expanded path names. The framework
only replaced the unexpanded pathname usages in the output that matched the environment
variable gtm_exe (which in turn matched the env var ydb_dist) with ##SOURCE_PATH##. For
the expanded pathname usages, it was expecting an env var gtm_exe_realpath to be set
coming in to the awk script. This was not being set in a YottaDB environment. So inref.awk
was enhanced to set this inside the awk script and replace these usages with ##SOURCE_REALPATH##.

With this framework level change, the v63003/gtm8735 subtest reference file was fixed to
use ##SOURCE_REALPATH## wherever it had the expanded path usages.