Fix occasional v63004/gtm8909 subtest failure due to extra MUPIP> or DSE> prompt
We had two failures till now on slower systems (armv6l) and each time it was
an extra MUPIP> or DSE> prompt.

We are sending two \r in the expect script. Not sure why that is necessary.
That is the cause of the extra MUPIP> or DSE> prompt.

Have therefore reworked the test to do a better job with just one \r.
Have run this test on all boxes (x86_64, armv7l and armv6l) 10 times and it passes fine.