Fix occasional v63004/gtm8791 subtest failure due to LKE> prompt showing up after the fg command
When the backgrounded lke process is foregrounded by the "fg" command, the shell prints the
following most of the times.

/usr/library/V999_R122/dbg/lke
<blank line>

With the LKE process waiting for input at the blank line.
But sometimes the shell prints the following instead.

/usr/library/V999_R122/dbg/lke
LKE>

With the LKE process waiting for input after the "LKE> " prompt in the second line.
It is not yet clear why this difference exists but because of this the v63004/gtm8791 subtest
fails. In case there is a LKE> prompt after the fg command, because the expect script sends a
"\r" (which causes another LKE> prompt), there is a total of 2 LKE> prompts and the expect
script incorrectly matches the first LKE> prompt and after sending "exit\r", it incorrectly
matches ">" (the shell prompt) with the second "LKE>" prompt.

This is now fixed by having the shell prompt to not be a substring of the LKE prompt.