Fix very rare ideminter_rolrec/interrupted_rollback_or_recover subtest failure due to a timing issue
We had an in-house test failure with the following diff

./ideminter_rolrec_0_5/interrupted_rollback_or_recover/close_reorg_upgrd_dwngrd.out
TEST-E-MUPIP is not executing!

In this case, the com/mu_reorg_upgrd_dwngrd_crash.csh script waits for 60 seconds,
checking every 1 second if the mupip reorg process can be found. And it found that
process in the 60th (last) iteration and was able to kill it but still ended up issuing
the TEST-E-MUPIP error. Therefore, the timeout is now increased to 300 seconds.

This opportunity is used to further enhance logging in the script. Having the output of
multiple lines in a script go to different log files (${KILL_LOG} and ${KILL_LOG}_ps)
makes it very hard to analyze. Also record the time as part of every ps output.