Fix suppl_inst_A_1/instinfoprop subtest due to lack of space in /tmp for temporary files created by <sort>
A prior commit (SHA 3c5e37562cad7482a438aaad3c006336ed713c08) fixed sort usages to be $sort
where the $sort variable value was "sort -T" to redirect the temporary file directory to be
the test output directory (instead of /tmp).

But it missed out a couple of sort usages which were at the end of a long line. And we had
a test failure due to one such sort in an internal test. Those couple usages are now fixed.