Fix occasional v62000/gtm7919 subtest failure due to getoper.csh usage not capturing all syslog messages till then
getoper.csh captures all messages since a particular time until the instant the getoper.csh
script is called correctly if the "" parameter is passed as $2 ($time_after). It does this
by internally generating a random string and sending it to the syslog and waiting for that
to show up in the syslog at which point it is guaranteed all pending syslog messages till
then will show up in the syslog before the random string.

The alternative usage of supplying a non-null $2 ($time_after) is that we might miss out
on syslog messages that are asynchronously being delivered but not yet recorded in the
journalctl logs. This is the current usage and could cause occasional failures.