Fix set_jnl_1/autoswitchtp subtest failure due to update process being killed by OOM killer
The update process was killed by the OOM (out-of-memory) killer on a 1-CPU machine with just
512Mb total memory. This caused the test to hang/fail.

This test creates 9 regions each with 8K global buffers and 4K blocksize. That itself accounts
for a total of at least 288Mb memory. And jnlpool/recvpool etc. will take the total to > 300Mb.

To reduce the memory requirements of the update process, the dbcreate now sets the global
buffers to 1K instead of 8K. This should reduce the total memory to < 100Mb which should
hopefully avoid the OOM killer from kicking in.