Enhance dbcreate/dbcheck to work with subdirectories holding .gld/.dat files
dbcreate.csh currently only creates gld and db files in the local and remote sides of the
test output directory. It does not let the user create the gld/db in a subdirectory on the
local/remote side of the test. That need is now implemented in dbcreate and dbcheck.

The syntax currently to create a mumps.gld and mumps.dat/a.dat
	dbcreate.csh mumps 2

The new syntax to create a mumps.gld and mumps.dat/a.dat under the dir1 directory is
	dbcreate.csh dir1/mumps 2

i.e. instead of "mumps", one needs to specify "dir1/mumps".

Similarly, the syntax currently to check the mumps.dat/a.dat db is
	dbcheck.csh

The new syntax to dbcheck mumps.dat/a.dat under the dir1 subdirectory is
	dbcheck.csh dir1/mumps

The new syntax causes both dbcreate and dbcheck to go into dir1 and see if a mumps.gld file
exists (because dir1/mumps was specified) and if so it will use that as gtmgbldir/ydb_gbldir
and do the mupip create or mupip integ -reg "*" respectively.